Moves EntityCard, InfoButton, Inputs and PopularTables
