+++ amundsen_application/static/js/components/Inputs/CheckBoxItem/index.spec.tsx
