+++ amundsen_application/static/js/components/Inputs/CheckBoxItem/index.tsx
@@ -2,9 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import * as React from 'react';
-import { connect } from 'react-redux';
 
-import 'components/common/Inputs/styles.scss';
+import 'components/Inputs/styles.scss';
 
 export interface CheckBoxItemProps {
   checked?: boolean;
