+++ amundsen_application/static/js/components/common/PopularTables/index.spec.tsx
@@ -4,7 +4,7 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import InfoButton from 'components/common/InfoButton';
+import InfoButton from 'components/InfoButton';
 import PaginatedResourceList from 'components/common/ResourceList/PaginatedResourceList';
 import globalState from 'fixtures/globalState';
 import {
