+++ amundsen_application/static/js/components/common/PopularTables/index.tsx
@@ -5,7 +5,7 @@ import * as React from 'react';
 
 import { TableResource } from 'interfaces';
 
-import InfoButton from 'components/common/InfoButton';
+import InfoButton from 'components/InfoButton';
 import PaginatedResourceList from 'components/common/ResourceList/PaginatedResourceList';
 import ShimmeringResourceLoader from 'components/common/ShimmeringResourceLoader';
