+++ amundsen_application/static/js/pages/SearchPage/SearchFilter/CheckBoxFilter/index.spec.tsx
@@ -4,13 +4,13 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
 
-import CheckBoxItem from 'components/common/Inputs/CheckBoxItem';
+import CheckBoxItem from 'components/Inputs/CheckBoxItem';
 
 import globalState from 'fixtures/globalState';
 
 import { GlobalState } from 'ducks/rootReducer';
 
-import { FilterType, ResourceType } from 'interfaces';
+import { ResourceType } from 'interfaces';
 import {
   CheckBoxFilter,
   CheckBoxFilterProps,
