+++ amundsen_application/static/js/pages/SearchPage/SearchFilter/FilterSection/index.tsx
@@ -12,7 +12,7 @@ import {
 
 import { GlobalState } from 'ducks/rootReducer';
 import { FilterType, IconSizes } from 'interfaces';
-import InfoButton from 'components/common/InfoButton';
+import InfoButton from 'components/InfoButton';
 import { CLEAR_BTN_TEXT } from '../constants';
 
 import CheckBoxFilter, { CheckboxFilterProperties } from '../CheckBoxFilter';
